refactor: continue replacing adapter pattern usage with direct CRUD operations
- Update entity_resolver.py to use CRUD operations directly
- Update trend_detector.py to use engine.with_session instead of adapter.with_session
- Update historical_aggregator.py to use engine.with_session
- Update topic_analyzer.py to use engine.with_session
- Update sentiment_analyzer.py to use direct CRUD operations
- Update opinion_visualizer.py to use engine.with_session
- Update headline_trend_flow.py to use engine.with_session
